+++ poem/src/test/mod.rs
@@ -1,11 +1,13 @@
 //! Test utilities to test your endpoints.
 
 mod client;
+mod form;
 mod json;
 mod request_builder;
 mod response;
 
 pub use client::TestClient;
+pub use form::{TestForm, TestFormField};
 pub use json::{TestJson, TestJsonArray, TestJsonObject, TestJsonValue};
 pub use request_builder::TestRequestBuilder;
 pub use response::TestResponse;